+++ lib/src/type_helper.dart
@@ -10,16 +10,10 @@ import 'package:source_gen/source_gen.dart' show TypeChecker;
 abstract class TypeHelper {
   const TypeHelper();
 
-  /// Returns `true` if this can serialize a value of type [targetType] into a
-  /// JSON literal.
-  bool canSerialize(DartType targetType);
-
-  /// Returns `true` if this can deserialize a JSON literal into a value of
-  /// type [targetType].
-  bool canDeserialize(DartType targetType);
-
   /// Returns Dart code that serializes an [expression] representing a Dart
   /// object of type [targetType].
+  ///
+  /// If [targetType] is not supported, returns `null`.
   ///
   /// Let's say you want to serialize a class `Foo` as just its `id` property
   /// of type `int`.
@@ -35,6 +29,8 @@ abstract class TypeHelper {
 
   /// Returns Dart code that deserializes an [expression] representing a JSON
   /// literal to into [targetType].
+  ///
+  /// If [targetType] is not supported, returns `null`.
   ///
   /// Let's say you want to deserialize a class `Foo` by taking an `int` stored
   /// in a JSON literal and calling the `Foo.fromInt` constructor.
@@ -60,15 +56,7 @@ abstract class TypeHelper {
 class JsonHelper extends TypeHelper {
   const JsonHelper();
 
-  //TODO(kevmoo): This should be checking for toJson method, but toJson might be
-  //   gone during generation, so we'll have to check for the annotation, too!
-  // In the mean time, just assume the `canSerialize` logic will work most of
-  //   the time.
-  @override
-  bool canSerialize(DartType type) => canDeserialize(type);
-
-  @override
-  bool canDeserialize(DartType type) {
+  bool _canDeserialize(DartType type) {
     if (type is! InterfaceType) return false;
 
     var classElement = type.element as ClassElement;
@@ -88,10 +76,24 @@ class JsonHelper extends TypeHelper {
   /// By default, JSON encoding in from `dart:convert` calls `toJson` on
   /// provided objects.
   @override
-  String serialize(DartType targetType, String expression) => expression;
+  String serialize(DartType targetType, String expression) {
+    // TODO(kevmoo): This should be checking for toJson method, but toJson might
+    //   be gone during generation, so we'll have to check for the annotation.
+    // In the mean time, just assume the `canSerialize` logic will work most of
+    //   the time.
+    if (!_canDeserialize(targetType)) {
+      return null;
+    }
+
+    return expression;
+  }
 
   @override
   String deserialize(DartType targetType, String expression) {
+    if (!_canDeserialize(targetType)) {
+      return null;
+    }
+
     // TODO: the type could be imported from a library with a prefix!
     // github.com/dart-lang/json_serializable/issues/19
     return "new ${targetType.name}.fromJson($expression as Map<String, dynamic>)";
@@ -101,22 +103,19 @@ class JsonHelper extends TypeHelper {
 class DateTimeHelper extends TypeHelper {
   const DateTimeHelper();
 
-  @override
-  bool canSerialize(DartType type) => _matchesType(type);
-
-  @override
-  bool canDeserialize(DartType type) => _matchesType(type);
-
   bool _matchesType(DartType type) =>
       const TypeChecker.fromUrl('dart:core#DateTime').isExactlyType(type);
 
   @override
   String serialize(DartType targetType, String expression) =>
-      "$expression?.toIso8601String()";
+      _matchesType(targetType) ? "$expression?.toIso8601String()" : null;
 
   @override
   String deserialize(DartType targetType, String expression) =>
-      // TODO(kevmoo) `String` here is ignoring
-      // github.com/dart-lang/json_serializable/issues/19
-      "DateTime.parse($expression as String)";
+      _matchesType(targetType)
+          ?
+          // TODO(kevmoo) `String` here is ignoring
+          // github.com/dart-lang/json_serializable/issues/19
+          "DateTime.parse($expression as String)"
+          : null;
 }